ログイン、ログアウト機能完成

+++ database/migrations/2020_05_03_070009_create_notes_table.php
@@ -19,7 +19,7 @@ class CreateNotesTable extends Migration
             $table->text('comment');
             $table->text('preview_image')->nullable();
             $table->string('url')->nullable();
-            $table->integer('evaluation')->nullable();
+            $table->float('evaluation',2,1)->nullable();
             $table->timestamps();
             $table->softDeletes();
         });
